adds token auth and testing

+++ controllers/blogs.js
@@ -1,23 +1,32 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
 
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0
-  })
-
   try {
+    console.log('decoded token', request.decodedToken)
+    const user = await User.findById(request.decodedToken.id)
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user._id
+    })
+
     const savedBlog = await blog.save()
+    console.log(savedBlog)
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.json(savedBlog.toJSON())
   } catch (exception) {
     next(exception)
@@ -46,9 +55,17 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    console.log('params', request.decodedToken)
+    console.log('blog', blog)
+
+    if (blog.user.toString() === request.decodedToken.id) {
+      console.log('DELETING')
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
   } catch (exception) {
+    console.log('blog user and user id do not match')
     next(exception)
   }
 })
